Squashed commit of the following:
commit b95834c772143b486c1a4159747334d8b70d6d3d
Author: Christopher Seven Phiri <chris7phiri@gmail.com>
Date:   Wed Dec 15 00:20:42 2021 +0200

    added flextheme

commit cd0415ad0fa1bff5d5bccab4b2ecf08c913d0246
Author: Christopher Seven Phiri <chris7phiri@gmail.com>
Date:   Tue Dec 14 23:56:40 2021 +0200

    moved settings widgets

commit 906293ee0ec070d213912a93539cdb35afc2bfc0
Author: Christopher Seven Phiri <chris7phiri@gmail.com>
Date:   Tue Dec 14 00:55:30 2021 +0200

    removed unused auth code

commit 8ec66bc84d13c9f23571e94330a6fdf3233947d5
Author: Christopher Seven Phiri <chris7phiri@gmail.com>
Date:   Tue Dec 14 00:42:00 2021 +0200

    added fire auth ui
