--- lib/src/blocs/register/register_cubic.dart
@@ -1,25 +1,0 @@
-import 'package:bloc/bloc.dart';
-import 'package:equatable/equatable.dart';
-import 'package:formz/formz.dart';
-
-import '../../repositories/auth_repository/authentication_repository.dart';
-
-part 'register_state.dart';
-
-class RegisterCubit extends Cubit<RegisterState> {
-  RegisterCubit(this._authenticationRepository) : super(const RegisterState());
-
-  final AuthenticationRepository _authenticationRepository;
-
-  Future<void> logInWithGoogle() async {
-    emit(state.copyWith(status: FormzStatus.submissionInProgress));
-    try {
-      await _authenticationRepository.logInWithGoogle();
-      emit(state.copyWith(status: FormzStatus.submissionSuccess));
-    } on Exception {
-      emit(state.copyWith(status: FormzStatus.submissionFailure));
-    } on NoSuchMethodError {
-      emit(state.copyWith(status: FormzStatus.pure));
-    }
-  }
-}
