--- lib/src/screens/login.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-
-import '../widgets/login_form.dart';
-import '../blocs/login/login_cubit.dart';
-import '../repositories/auth_repository/authentication_repository.dart';
-
-class LoginDialog extends StatelessWidget {
-  const LoginDialog({Key? key}) : super(key: key);
-
-  static Dialog dialog() => const Dialog(child: LoginDialog());
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 400,
-      padding: const EdgeInsets.all(8.0),
-      child: BlocProvider(
-        create: (_) => LoginCubit(context.read<AuthenticationRepository>()),
-        child: const LoginForm(),
-      ),
-    );
-  }
-}
