--- lib/src/screens/welcome.dart
@@ -1,48 +1,0 @@
-import 'package:firebase_bloc_starter/AppKeys.dart';
-import 'package:firebase_bloc_starter/src/screens/login.dart';
-import 'package:firebase_bloc_starter/src/screens/register.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-class WelcomePage extends StatelessWidget {
-  const WelcomePage({Key? key}) : super(key: key);
-  static Page page() => const MaterialPage<void>(child: WelcomePage());
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Welcome')),
-      body: Container(
-        padding: EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            ElevatedButton(
-              key: AppKeys.LoginButtonKey,
-              onPressed: () {
-                showDialog(
-                    context: context,
-                    builder: (context) {
-                      return LoginDialog.dialog();
-                    });
-              },
-              child: Text("Login"),
-            ),
-            ElevatedButton(
-              key: AppKeys.RegisterButtonKey,
-              onPressed: () {
-                showDialog(
-                    context: context,
-                    builder: (context) {
-                      return RegisterDialog.dialog();
-                    });
-              },
-              child: Text("SignUp"),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
