--- lib/src/widgets/login_form.dart
@@ -1,49 +1,0 @@
-import 'package:firebase_bloc_starter/AppKeys.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter/material.dart';
-import 'package:formz/formz.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import '../blocs/login/login_cubit.dart';
-import 'google_button.dart';
-
-class LoginForm extends StatelessWidget {
-  const LoginForm({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocListener<LoginCubit, LoginState>(
-      listener: (context, state) {
-        if (state.status.isSubmissionFailure) {
-          ScaffoldMessenger.of(context)
-            ..hideCurrentSnackBar()
-            ..showSnackBar(
-              const SnackBar(content: Text('Authentication Failure')),
-            );
-        }
-      },
-      child: Align(
-        alignment: const Alignment(0, -1 / 3),
-        child: SingleChildScrollView(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                AppLocalizations.of(context)!.login,
-                style: Theme.of(context).textTheme.headline5,
-              ),
-              const SizedBox(height: 16.0),
-              GoogleLoginButton(
-                key: AppKeys.GoogleButtonKey,
-                onPressed: () {
-                  context.read<LoginCubit>().logInWithGoogle();
-                  Navigator.of(context).pop();
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
