+++ pubspec.yaml
@@ -40,14 +40,14 @@ dependencies:
   very_good_analysis: ^2.1.0
   cupertino_icons: ^1.0.2
   uuid: ^3.0.4
-  cloud_firestore: "^2.4.0"
+  cloud_firestore: ^3.1.4
   fake_cloud_firestore: ^1.1.2
   google_sign_in_mocks: ^0.2.1
   firebase_auth_mocks: ^0.8.0
   flutter_local_notifications: ^8.1.1+1
   http: ^0.13.3
   mixpanel_flutter: ^1.3.1
-  flutter_svg: ^0.23.0+1
+  flutter_svg: ^0.22.0
   url_launcher: ^6.0.10
   google_fonts: ^2.1.0
   shared_preferences: ^2.0.7
@@ -56,6 +56,8 @@ dependencies:
   sign_in_with_apple: ^3.2.0
   crypto: ^3.0.1
   flutter_dotenv: ^5.0.2
+  flutterfire_ui: ^0.2.0+2
+  flex_color_scheme: ^4.1.1
 
 dev_dependencies:
   flutter_test:
@@ -79,7 +81,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - .env
-  #   - images/a_dot_burr.jpeg
+    - assets/images/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
